requirements.txt
altair==4.2.0
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
Pillow==9.2.0
streamlit==1.10.0
